+++ addons/missionModules/functions/fnc_zeus_missionHint.sqf
@@ -1,3 +1,5 @@
 #include "script_component.hpp"
 
 TRACE_1("params",_this);
+
+//Everything handled by GUI
